+++ ios/PrayerApp/PrayerWidget.m
@@ -65,4 +65,29 @@ RCT_EXPORT_METHOD(setWidgetHighlightDynamic:(BOOL)enabled
   resolve(nil);
 }
 
+RCT_EXPORT_METHOD(setIosWidgetHighlightAppearance:(NSString *)highlightId
+                  hex:(NSString *)hex
+                  dynamic:(BOOL)dynamic
+                  resolver:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+{
+  NSUserDefaults *group =
+      [[NSUserDefaults alloc] initWithSuiteName:@"group.com.prayerapp"];
+  NSUserDefaults *target = group != nil ? group : [NSUserDefaults standardUserDefaults];
+  NSString *hid =
+      (highlightId != nil && highlightId.length > 0) ? highlightId : @"green";
+  [target setObject:hid forKey:@"widget_highlight_id"];
+  if (hex != nil && hex.length > 0) {
+    [target setObject:hex forKey:@"widget_highlight_hex"];
+  } else {
+    [target removeObjectForKey:@"widget_highlight_hex"];
+  }
+  [target setBool:dynamic forKey:@"widget_highlight_dynamic"];
+  [target synchronize];
+  if (@available(iOS 14.0, *)) {
+    [[WidgetCenter sharedWidgetCenter] reloadAllTimelines];
+  }
+  resolve(nil);
+}
+
 @end